learn laravel basic

+++ routes/web.php
@@ -3,8 +3,11 @@
 use App\Http\Controllers\StudentController;
 use App\Http\Controllers\SinhvienController;
 use App\Http\Controllers\CourseController;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\DatatablesController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Middleware\checkLoginMiddleware;
+use App\Http\Middleware\checkSuperAdminMiddleware;
 use App\Http\Controllers\TestController;
 
 /*
@@ -18,42 +21,33 @@ use App\Http\Controllers\TestController;
 |
 */
 
-Route::resource('student', StudentController::class)->except('show');
+Route::get('login', [AuthController::class , 'login'])->name('login');
+Route::post('login', [AuthController::class , 'process_login'])->name('process_login');
 
+Route::group(['middleware' => checkLoginMiddleware::class], function () {
 
-// Route::group(['prefix' => 'student', 'as' => 'student.'], function () {
-//     Route::get('/', [StudentController::class, 'index'])->name('index');
-//     Route::get('/create', [StudentController::class, 'create'])->name('create');
-//     Route::post('/create', [StudentController::class, 'store'])->name('store');
-//     Route::delete('/destroy/{student}', [StudentController::class, 'destroy'])->name('destroy');
-//     Route::get('/edit/{student}', [StudentController::class, 'edit'])->name('edit');
-//     Route::put('/edit/{student}', [StudentController::class, 'update'])->name('update');
-// });
+    Route::get('/logout', [AuthController::class, 'logout'])->name('logout');
 
-Route::group(['prefix' => 'courses', 'as' => 'courses.'], function () {
-    Route::get('/', [CourseController::class, 'index'])->name('index');
-    Route::get('/create', [CourseController::class, 'create'])->name('create');
-    Route::post('/create', [CourseController::class, 'store'])->name('store');
-    Route::delete('/destroy/{courses}', [CourseController::class, 'destroy'])->name('destroy');
-    Route::get('/edit/{courses}', [CourseController::class, 'edit'])->name('edit');
-    Route::put('/edit/{courses}', [CourseController::class, 'update'])->name('update');
-});
+    Route::resource('student', StudentController::class)->except('show');
 
-Route::group(['prefix' => 'sinhvien', 'as' => 'sinhvien.'], function () {
-    Route::get('/', [SinhvienController::class, 'index'])->name('index');
-    Route::get('/create', [SinhvienController::class, 'create'])->name('create');
-    Route::post('/create', [SinhvienController::class, 'store'])->name('store');
-    Route::delete('/destroy/{sinhvien}', [SinhvienController::class, 'destroy'])->name('destroy');
-    Route::get('/edit/{sinhvien}', [SinhvienController::class, 'edit'])->name('edit');
-    Route::put('/edit/{sinhvien}', [SinhvienController::class, 'update'])->name('update');
-});
+    Route::group(['prefix' => 'courses', 'as' => 'courses.'], function () {
+        Route::get('/', [CourseController::class, 'index'])->name('index');
+        Route::get('/create', [CourseController::class, 'create'])->name('create');
+        Route::post('/create', [CourseController::class, 'store'])->name('store');
+        Route::delete('/destroy/{courses}', [CourseController::class, 'destroy'])->name('destroy')->middleware(checkSuperAdminMiddleware::class);
+        Route::get('/edit/{courses}', [CourseController::class, 'edit'])->name('edit');
+        Route::put('/edit/{courses}', [CourseController::class, 'update'])->name('update');
+    });
 
-Route::get('/datatables', [DatatablesController::class, 'getIndex'])->name('datatables');
-Route::get('/datatables/data', [DatatablesController::class, 'anyData'])->name('datatables.data');
+    Route::group(['prefix' => 'sinhvien', 'as' => 'sinhvien.'], function () {
+        Route::get('/', [SinhvienController::class, 'index'])->name('index');
+        Route::get('/create', [SinhvienController::class, 'create'])->name('create');
+        Route::post('/create', [SinhvienController::class, 'store'])->name('store');
+        Route::delete('/destroy/{sinhvien}', [SinhvienController::class, 'destroy'])->name('destroy')->middleware(checkSuperAdminMiddleware::class);
+        Route::get('/edit/{sinhvien}', [SinhvienController::class, 'edit'])->name('edit');
+        Route::put('/edit/{sinhvien}', [SinhvienController::class, 'update'])->name('update');
+    });
 
-
-Route::get('/courses/api' , [CourseController::class, 'api'])->name('courses.api');
-
-Route::get('test', function () {
-    return view('layout.master');
+    Route::get('/datatables', [DatatablesController::class, 'getIndex'])->name('datatables');
+    Route::get('/datatables/data', [DatatablesController::class, 'anyData'])->name('datatables.data');
 });
